Message관련 Resolver 기능별로 Component화
