+++ src/components/main/messages/DeleteRoom.js
@@ -1,7 +1,7 @@
 /* 
 작성자 : SJ
 작성일 : 2022.01.15
-수정일 : ------
+수정일 : 2022.01.16
 */
 
 /*
