+++ src/screens/main/messages/SeeRoom.js
@@ -1,7 +1,7 @@
 /* 
 작성자 : SJ
 작성일 : 2022.01.14
-수정일 : 2022.01.15
+수정일 : 2022.01.16
 */
 
 /* 
@@ -13,25 +13,14 @@
 */
 
 import { gql, useApolloClient, useMutation, useQuery, useReactiveVar } from '@apollo/client'
-import { faArrowCircleUp } from '@fortawesome/free-solid-svg-icons'
 import { useEffect, useState } from 'react'
-import { useForm } from 'react-hook-form'
-import styled from 'styled-components'
 import useLoggedInUser from '../../../hooks/useLoggedInUser'
-import { darkModeVar, chatRoomIdVar, chatUserIdVar } from '../../../utils/apollo'
-import { colors } from '../../../utils/styles'
-import Input from '../../../components/shared/form/Input'
-import InputWithFontAwesome from '../../../components/shared/form/InputWithFontAwesome'
+import { chatRoomIdVar, chatUserIdVar } from '../../../utils/apollo'
 import { MESSAGE_DEFAULT_FRAGMENT } from '../../../components/shared/utils/fragments'
 import Message from '../../../components/main/messages/Message'
 import MessageRoomLayout from '../../../components/layouts/MessageRoomLayout'
+import CreateMessage from '../../../components/main/messages/CreateMessage'
 
-const PositionAbsolute = styled.div`
-    position:absolute;
-    bottom:2px;
-    width:100%;
-    margin: 0 auto;
-`
 const SEE_ROOM = gql`
     query seeRoom($roomId:Int,$userId:Int,$offset:Int){
         seeRoom(roomId:$roomId,userId:$userId){
@@ -58,15 +47,6 @@ const READ_MESSAGES = gql`
     }
 `
 
-const CREATE_MESSAGE = gql`
-    mutation createMessage($payload:String!,$roomId:Int,$userId:Int){
-        createMessage(payload:$payload,roomId:$roomId,userId:$userId){
-            ...MessageDefaultFragment
-        }
-    }
-    ${MESSAGE_DEFAULT_FRAGMENT}
-`
-
 const REALTIME_ROOM = gql`
     subscription realtimeRoom($id:Int!){
         realtimeRoom(id:$id){
@@ -77,12 +57,9 @@ const REALTIME_ROOM = gql`
 `
 
 export default function SeeRoom() {
-    const darkMode = useReactiveVar(darkModeVar)
     const chatRoomId = useReactiveVar(chatRoomIdVar)
     const chatUserId = useReactiveVar(chatUserIdVar)
     const [messageData, setMessageData] = useState([])
-
-    const { register, handleSubmit, setValue } = useForm()
 
     // readMessages Mutation 구현부
     const updateReadMessages = (cache, { data }) => {
@@ -149,36 +126,6 @@ export default function SeeRoom() {
             })
         }
     }
-    const updateCreateMessage = (cache, { data: { createMessage } }) => {
-        if (createMessage.id) {
-            cache.modify({
-                id: `Room:${data?.seeRoom?.id}`,
-                fields: {
-                    messages(prev) {
-                        return [...prev, createMessage]
-                    }
-                }
-            })
-            setValue("message", "")
-        }
-    }
-
-
-    // createMessage 구현부
-    const [createMessage, { loading: sendLoading }] = useMutation(CREATE_MESSAGE, {
-        update: updateCreateMessage
-    })
-    const onValid = ({ message }) => {
-        if (sendLoading) return;
-
-        createMessage({
-            variables: {
-                payload: message,
-                ...(chatRoomId && { roomId: parseInt(chatRoomId) }),
-                ...(chatUserId && { userId: parseInt(chatUserId) }),
-            }
-        })
-    }
 
     /* 
     1. seeRoom Query가 로드되면 Messages들을 messageData state에 옮겨 담는다.
@@ -231,21 +178,7 @@ export default function SeeRoom() {
                     message={message.payload}
                 />
             )}
-            <PositionAbsolute>
-                <InputWithFontAwesome
-                    icon={faArrowCircleUp}
-                    size={"lg"}
-                    onClick={handleSubmit(onValid)}
-                    color={darkMode ? colors.white : colors.black}
-                >
-                    <Input
-                        {...register("message", {
-                            required: true,
-                        })}
-                        required
-                    />
-                </InputWithFontAwesome>
-            </PositionAbsolute>
+            <CreateMessage roomId={data?.seeRoom?.id} />
         </MessageRoomLayout>
     )
 }
